fixed some console errors (missing scripts), use meetup API to pull upcoming meetups (test feed)
I'm pulling test upcoming meedups from a mammoth meetup group right now
(they have two upcoming feeds).  To do this I added a controller.js and
linked it before end of body and added controller UpcomingMeetups to the
partial view for highlights.  Everything appears to be working fine but
I need to figure out how to modify some of the JSONP data (i.e. time is
in milleseconds) -- need to do a new Date to the time for each event to
get the actual date but I'll figure that out later.